+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
     "python.testing.pytestArgs": [
-        "qa"
+        "testing"
     ],
     "python.testing.unittestEnabled": false,
     "python.testing.pytestEnabled": true
