+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-testpaths=qa
+testpaths=testing
 asyncio_mode=auto
 python_files=*.py
 log_file=test.log
